[Oztechan/CCC#2611] Move MainView to MainRootView
